Sauvegarde des temps dans bdd
-> Sauvegarde de tous les temps de chaque participants pour chaque étape dans la bdd

+++ app/src/main/java/com/example/gestion_course/entities/Temps.kt
@@ -22,5 +22,5 @@ import androidx.room.Index
 data class Temps (
         val num_etape_temps: Int,
         val num_participant_temps: Int,
-        val time: Long
+        val time: Int
 )
